# lib/api-zod/src/generated/types/adminLeadList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminLeadRow } from "./adminLeadRow";

export interface AdminLeadList {
  leads: AdminLeadRow[];
}