# lib/api-zod/src/generated/types/adminLeadRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminLeadRowType } from "./adminLeadRowType";

export interface AdminLeadRow {
  id: number;
  type: AdminLeadRowType;
  name: string;
  email: string;
  org: string;
  locale?: string | null;
  city?: string | null;
  message?: string | null;
  createdAt: Date;
}