# lib/api-zod/src/generated/types/chatBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatBodyLocation } from "./chatBodyLocation";
import type { ChatBodyProfile } from "./chatBodyProfile";
import type { ChatMessage } from "./chatMessage";

export interface ChatBody {
  /** @maxLength 64 */
  sessionId?: string;
  /** @maxLength 8 */
  locale: string;
  /**
   * @minItems 1
   * @maxItems 16
   */
  messages: ChatMessage[];
  location?: ChatBodyLocation;
  profile?: ChatBodyProfile;
}