# lib/api-zod/src/generated/types/conditions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConditionsAirRisk } from "./conditionsAirRisk";
import type { ConditionsHeatRisk } from "./conditionsHeatRisk";

export interface Conditions {
  lat: number;
  lon: number;
  city?: string | null;
  country?: string | null;
  tempC?: number | null;
  feelsLikeC?: number | null;
  humidity?: number | null;
  windKph?: number | null;
  uvIndex?: number | null;
  aqiUs?: number | null;
  pm25?: number | null;
  summary: string;
  heatRisk: ConditionsHeatRisk;
  airRisk: ConditionsAirRisk;
  observedAt: Date;
}