# lib/api-zod/src/generated/types/pharmaLeadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PharmaLeadBody {
  /**
   * @minLength 1
   * @maxLength 200
   */
  name: string;
  /**
   * @minLength 1
   * @maxLength 200
   */
  company: string;
  /** @maxLength 200 */
  therapeutic?: string | null;
  /** @maxLength 320 */
  email: string;
  /** @maxLength 4000 */
  message?: string | null;
  /** @maxLength 8 */
  locale?: string | null;
  /** @maxLength 1024 */
  pageUrl?: string | null;
}